### **User description**
#123404

___

### **PR Type**
Bug fix

___

### **Description**
- Corrected the data type of `PAYORDNO` column.

___

### Diagram Walkthrough

```mermaid
flowchart LR
  A["Examine PAYORDNO column type"] --> B{"Incorrect data type?"}
  B -- "Yes" --> C["Change PAYORDNO type to numeric"]
  C --> D["Update AmitalContext template"]
  B -- "No" --> E["No changes needed"]
```

<details> <summary><h3> File Walkthrough</h3></summary>

<table><thead><tr><th></th><th align="left">Relevant files</th></tr></thead><tbody><tr><td><strong>Bug fix</strong></td><td><table>
<tr>
  <td>
    <details>
      <summary><strong>AmitalContext.Template.cs</strong><dd><code>Corrected `PAYORDNO` data type in `CCUPAYLINEF` entity</code>&nbsp; &nbsp; &nbsp; </dd></summary>
<hr>

/Logitude/Unifreight.Data/AmitalModel/AmitalContext.Template.cs

<ul><li>Changed the data type of the <code>PAYORDNO</code> property in the <code>CCUPAYLINEF</code> <br>entity from <code>int</code> to <code>numeric</code>.</ul>

</details>

  </td>
  <td><a href="https://amitaldata.visualstudio.com/Amital/_git/log-repo/pullrequest/6385?_a=files&path=/Logitude/Unifreight.Data/AmitalModel/AmitalContext.Template.cs">+4/-4</a>&nbsp; &nbsp; &nbsp; </td>

</tr>
</table></td></tr></tr></tbody></table>

</details>

___

Related work items: #123404

+++ Logitude/Unifreight.Data/AmitalModel/AmitalContext.Template.cs
@@ -12351,9 +12351,9 @@ namespace Unifreight.Data.AmitalModel
                     .HasDatabaseGeneratedOption(System.ComponentModel.DataAnnotations.Schema.DatabaseGeneratedOption.None)
                     .HasColumnType("decimal");
             modelBuilder.Entity<CCUPAYLINEF>()
-                .Property(p => p.PAYORDNO)
-                    .HasColumnName(@"PAY_ORD_NO")
-                    .HasColumnType("int");
+    .Property(p => p.PAYORDNO)
+    .HasColumnName(@"PAY_ORD_NO")
+    .HasColumnType("numeric");
             modelBuilder.Entity<CCUPAYLINEF>()
                 .Property(p => p.PAYMETHOD)
                     .HasColumnName(@"PAY_METHOD")
